Remove redundancy between AnimatedAdapter and OpCalculator
animateDataSetChanged() can be used to animate changes on the UI thread.
prepareDataSetChanges() can be used on a background thread, followed by
a call to animatePendingDataSetChanged() on the UI thread for large or
complex data sets.

Also renames getChangeId() to getChangeHash(), improving the intent of
the method. This is not an identifier, but instead a hash that
represents the content for the given animation id. Multiple items can
have the same change hash.