Remove sticky header immediately if the entry it represents is removed
Prevents crashing when the layout that follows occurs:
java.lang.IndexOutOfBoundsException:
Inconsistency detected. Invalid item position 0(offset:-1).state:8

Having the sticky header view attached by the time measuring / laying
out starts messes up the position calculation for the animations. By
doing this we effectively account for this behaviour, with the downside
being that we don’t have a chance to recycle the view.