Fix AsyncDataSetDiffer workflow with adapter observing
AsyncDataSetDiffer does not run linearly:
1. It gets a call to `diffDataSet` on the UI thread
2. It calculates the differences on a background thread
3. It propagates the changes in a UI thread

In step 1, the adapter stops being observed to avoid double-counting
changes. In step 3, it starts being observed again.

Multiple calls to `diffDataSet` can start without the first finishing,
which triggers unwanted calls to `stopObservingItems` without the
corresponding calls to `startObservingItems`.

Adding a counter ensures the adapter stops being observed on the first
call, and only starts observing again in the last call, which is the
correct behavior.