Rely on touch position instead of view position when dragging
With the newly added flags, including the option to not translate the
dragged view as the drag&drop is ongoing, relying on its position on
screen is error prone. Particularly, it became problematic for
scrolling when the translation flag was disabled.

An indirect benefit of the revamped workflow is that drag & drop can now
be started for an item that's not yet laid out. This means users can
insert an item in the adapter, call notifyItemInserted, and start
drag&drop immediately afterwards, without having to wait for the next
layout pass.